requirements.txt
google-genai
sentence-transformers
faiss-cpu
numpy
gradio